aergoluac: save raw decoded payload to file
move packing of contract to luac
